trips/trips_manager/requirements.txt
pymongo==4.6.1
requests==2.31.0
kafka-python==2.0.2
geopy==2.4.1

